requirements.txt
Django
djangorestframework
gunicorn
pymongo
dnspython